Final Version Upload
Uploading the final version of Simon Says. This version handles holding
the top ten high scores. Adds images to the pieces, and the pieces can
be switched to different types of images. With a menu many options are
available to the user such as multiple sizes of game boards, changing
the difficulty of how the computer shows pieces.